pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.98.0",
    "openai>=1.60.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
    "tabulate>=0.9.0",
]
